+++ src/tokens/tokenize/hash.rs
